[UPD] - Update access modifiers to public for various functions and properties across multiple Kotlin files to enhance code clarity and maintain compliance with Explicit API mode.

+++ android/app/src/main/kotlin/com/jabook/app/jabook/audio/AudioPlayerServiceInitializer.kt
@@ -308,7 +308,7 @@ public class AudioPlayerServiceInitializer(
                         // Wait for controller with reasonable timeout for service initialization
                         val controller =
                             controllerFuture.get(
-                                com.jabook.app.jabook.audio.MediaControllerConstants.SERVICE_INIT_TIMEOUT_SECONDS,
+                                com.jabook.app.jabook.audio.MediaControllerConstants.SERVICE_INIT_TIMEOUT_SECONDS.toLong(),
                                 java.util.concurrent.TimeUnit.SECONDS,
                             )
                         service.serviceMediaController = controller
